chore: separated enums into their respective files
