+++ Intersect (Core)/Enums/StringVariableComparators.cs
@@ -1,0 +1,9 @@
+namespace Intersect.Enums
+{
+    public enum StringVariableComparators
+    {
+        Equal = 0,
+
+        Contains
+    }
+}
