+++ Intersect (Core)/Enums/VariableComparators.cs
@@ -1,0 +1,19 @@
+namespace Intersect.Enums
+{
+    public enum VariableComparators
+    {
+
+        Equal = 0,
+
+        GreaterOrEqual,
+
+        LesserOrEqual,
+
+        Greater,
+
+        Less,
+
+        NotEqual
+
+    }
+}
