+++ Intersect (Core)/Enums/VariableDataTypes.cs
@@ -1,0 +1,15 @@
+namespace Intersect.Enums
+{
+    public enum VariableDataTypes : byte
+    {
+
+        Boolean = 1,
+
+        Integer,
+
+        String,
+
+        Number
+
+    }
+}
